# billing-backend/src/main/java/io/github/logith/billing_application/request/RegisterRequest.java
package io.github.logith.billing_application.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank String name,
        @Email @NotBlank String email,
        @Size(min = 8) @NotBlank String password
) {

}